It Compiles!!!!
It compiles now. As well as added functionality to record the time
(currently just prints to screen and needs to be changed to out put the
time to the log file). It now also finds the .cpp file in the root
directory (though the root directory is just the current working
directory at the moment and needs to be changed to use argv[1] )