Got the program running so it test a program
Added function prototypes, edited code so that it works together.  Got
the main function pulling the .cpp file from command line and compiling.
Log file is opened and appended to.  Tested the program on a very simple
.cpp file but it does find the .tst files and steps into directories.
There was one bug with the final log write, if it passed all the tests
the output was 1E2% so there is a formatting issue with that.